YouTube-CommentsFetcher/requirements.txt
redis
python-dotenv
google-api-python-client